To run: create world (FlatLands), teleport to world (FlatLands), close and restart server. World should be implemented.
Updated project: Tree generator now fully implemented, random material
generator for material types implemented.

Idea of this was a swamp land, where parts of the ground are flooded.
Goal is to build a working island in the swamp and to mine all the
valuable minerals (mostly coal and iron, with diamonds and emeralds as
well).